chore: split out mock controller, pt 3
